[FEAT] modal: add new category
handle of add new category button. A modal will appear when the user
interacts with the button on the respective category type. In this modal,
will be possible to create/edit a category or a planning line.

A icon picker was added too, to be possible to show a icon related to the
category.