[REF] routes: folder architecture
The default remix routes will not be used anymore
in this project. Now, we'll define all routes
manually on the vite.config.ts.

This change is important, because we can customize
and easily split the dependencies for each route
in its folder.

doc: https://remix.run/docs/en/main/file-conventions/vite-config