[REF] monetary: Create monetary Input
Before this commit, to edit a monetary field, we was
doing it manually. Now, we can do it just using the
MonetaryInput component, that allow as to adjust the
signal, value and currency automatically, with the
same design developed for a Form View.

This commit applies the move code and fixes the correct
signature for all pages that are using monetary