[REF] dates helper: refactor dates helper & style
This commits adjusts some helpers for dates, now we have
the formatLongMonth when we want to use the name of the
month into a date, and formatSlashDivision, for dates
of type dd/mm/yyyy.

This commit applies a quick style tab for CategoriesSearch
and add the quickCreate option. This optional param can be
used to create or not a new record when searching on it.