[FIX] *: Change Date to string
Parsing Date format to string.

When a request is passed via loader to the client,
the dates with Date format was not being parsed
correctly, and it was delivering string types instead
of Date one. This commit fixes it, receving string in
all dates references and parsing to Date explicitly
to avoid access Date methods from a string.